+++ src/epicenter.ts
@@ -13,9 +13,9 @@ export {
     Router,
 } from 'utils/index';
 export {
-    Channel,
     accountAdapter,
     authAdapter,
+    // consensusAdapter,
     episodeAdapter,
     groupAdapter,
     leaderboardAdapter,
@@ -25,4 +25,5 @@ export {
     timeAdapter,
     vaultAdapter,
     worldAdapter,
+    Channel,
 } from 'adapters/index';